requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
defusedxml==0.7.1
Django==4.2.18
django-allauth==0.60.0
django-ckeditor==6.7.0
django-cors-headers==4.3.1
django-countries==7.5.1
django-crispy-forms==2.1
django-js-asset==2.2.0
django-resized==1.0.2
django-tinymce==3.6.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
filelock==3.13.1
fsspec==2024.2.0
idna==3.6
Jinja2==3.1.3
MarkupSafe==2.1.5
mpmath==1.3.0
mysqlclient
networkx==3.2.1
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.9.0.80
pillow==10.2.0
pycparser==2.21
PyJWT==2.8.0
PyMySQL==1.1.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
sympy==1.12
tensorflow_hub
torch==2.2.1
torchaudio==2.2.1
torchvision==0.17.1
typing_extensions==4.9.0
urllib3==2.1.0
whitenoise==6.6.0
